changed variable in main from float to double

+++ wine/net.cpp
@@ -281,7 +281,7 @@ int main()
 	ifstream ifs ("/data/winequality-red.csv");
 
 	char dummy;
-	float x;
+	double x;
 	for (int i = 0; i < 1; ++i) {
 		for (int i = 0; i < 11; ++i) {
 			ifs >> x;
